requirements.txt

---

## ⚙️ **requirements.txt**

```text
torch>=2.0.0
diffusers>=0.25.0
transformers>=4.36.0
accelerate>=0.27.0
safetensors>=0.4.0
gradio>=4.19.0
pillow>=10.0.0
numpy>=1.26.0
